+++ src/react/components/avatar/XUIAvatarCounter.js
@@ -6,6 +6,7 @@ import { sizeClassNames, classNames } from './constants';
 export default class XUIAvatarCounter extends PureComponent {
 	render() {
 		const { count, size, qaHook, className } = this.props;
+
 		const counterClassNames = cn(
 			classNames.base,
 			classNames.counter,
@@ -15,9 +16,9 @@ export default class XUIAvatarCounter extends PureComponent {
 
 		let value;
 
-		if(typeof count === 'string') {
+		if (typeof count === 'string') {
 			value = count;
-		} else if(count > 0) {
+		} else if (count > 0) {
 			value = '+' + count;
 		} else {
 			value = String(count);
@@ -32,10 +33,10 @@ XUIAvatarCounter.propTypes = {
 	className: PropTypes.string,
 
 	/** The count to display. If this is a string, it is passed through transparently. If it is a number, it will render with a + prefix */
-	count: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+	count: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
 
 	/** The size of the counter. Can be small, medium, large or xlarge */
-	size: PropTypes.oneOf(Object.keys(sizeClassNames))
+	size: PropTypes.oneOf(Object.keys(sizeClassNames)).isRequired
 };
 
 XUIAvatarCounter.defaultProps = {